requirements.txt
# Core Dependencies
numpy==1.26.3
pandas==2.1.4
pyarrow==14.0.2

# Data Sources
yfinance==0.2.36
newsapi-python==0.2.7
feedparser==6.0.10
praw==7.7.1
beautifulsoup4==4.12.3
requests==2.31.0

# Machine Learning & AI
torch==2.1.2
transformers==4.37.0
accelerate==0.25.0
huggingface-hub==0.20.3

# Sentiment Analysis
vaderSentiment==3.3.2

# Regime Detection & Statistics
hmmlearn==0.3.0
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.1

# Technical Analysis
pandas-ta==0.3.14b

# Storage & Caching
diskcache==5.6.3

# CLI & Visualization
rich==13.7.0
click==8.1.7
prompt-toolkit==3.0.43
plotext==5.2.8
tqdm==4.66.1

# Configuration & Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# Development & Testing
pytest==7.4.3
pytest-mock==3.12.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Optional but useful
jupyter==1.0.0
ipython==8.20.0
